Removed ui issues and code refactor
